% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{bandpower}
\alias{bandpower}
\title{Compute bandpower}
\usage{
bandpower(x)
}
\arguments{
\item{x}{The data}
}
\value{
A data frame of power in each band
}
\description{
Computes the bandpower on a default set
}
