% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{bp_pgram}
\alias{bp_pgram}
\title{Find the power within specified frequency bands.}
\usage{
bp_pgram(x, fs, freqs, psd = NULL, plot_pgram = FALSE)
}
\arguments{
\item{x}{The input signal}

\item{fs}{The signal sampling rate}

\item{freqs}{A vector or list of vectors of the
bandwidth's whose power should be computed}

\item{psd}{The output from stats::spec.pgram().}

\item{plot_pgram}{Plots default periodogram if TRUE}
}
\value{
A double or list of doubles representing the
          signal's frequency in the given bandwidths
}
\description{
Estimates the power within the given frequency 
 bands using Welch's method. The periogram is 
 then smoothed using wavelet thresholding. The 
 sum of power in bins within the frequency band
 is used as the estimate of that band's power.
}
