% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-features.R
\name{clean_feature}
\alias{clean_feature}
\title{Returns dataframe with feature(s)}
\usage{
clean_feature(feature)
}
\arguments{
\item{feature}{The feature to clean}
}
\description{
Returns dataframe with feature(s)
}

