% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ecomp_all}
\alias{ecomp_all}
\title{Compute epsilon complexity using bsplines}
\usage{
ecomp_all(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
Compute epsilon complexity using bsplines
}

