% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ecomp_cspline}
\alias{ecomp_cspline}
\title{Compute epsilon complexity using csplines}
\usage{
ecomp_cspline(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
Compute epsilon complexity using csplines
}

