% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ecomp_cspline_mae}
\alias{ecomp_cspline_mae}
\title{Compute epsilon complexity using csplines with 
mean absolute error.}
\usage{
ecomp_cspline_mae(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
Compute epsilon complexity using csplines with 
mean absolute error.
}

