% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ecomp_cspline_max}
\alias{ecomp_cspline_max}
\title{Compute epsilon complexity using csplines
 with max error.}
\usage{
ecomp_cspline_max(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
Compute epsilon complexity using csplines
 with max error.
}

