% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ecomp_cspline_mse}
\alias{ecomp_cspline_mse}
\title{Compute epsilon complexity using csplines with
 squared error.}
\usage{
ecomp_cspline_mse(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
Compute epsilon complexity using csplines with
 squared error.
}

