% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ecomp_cspline_rand}
\alias{ecomp_cspline_rand}
\title{Compute epsilon complexity using csplines
 and random sampling.}
\usage{
ecomp_cspline_rand(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
Compute epsilon complexity using csplines
 and random sampling.
}

