% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ecomp_lift}
\alias{ecomp_lift}
\title{Compute epsilon complexity using lifting}
\usage{
ecomp_lift(x)
}
\arguments{
\item{x}{The data}
}
\value{
The features
}
\description{
Compute epsilon complexity using lifting
}
