% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{fd_variogram}
\alias{fd_variogram}
\title{Variogram-based estimate of fractal dimension.}
\usage{
fd_variogram(x)
}
\arguments{
\item{x}{The data}
}
\value{
The results from fd.estimate
}
\description{
This function is a wrapper for the fd.estimate 
function of the fractaldim package.
}
