% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-features.R
\name{get_features}
\alias{get_features}
\title{Extract set of features from a time series.}
\usage{
get_features(data, features, id = NULL, ncores = NULL, verbose = FALSE)
}
\arguments{
\item{data}{A time series, matrix of dataframe.}

\item{features}{A list or vector of ts_features.}

\item{id}{Optional prefix for column headers.}

\item{ncores}{Parallelize feature extraction (Linux only).}

\item{verbose}{Print status to console.}
}
\value{
A data frame with row of features for each time.
 series in data.
}
\description{
Extract set of features from a time series.
}
