% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{permutation_entropy}
\alias{permutation_entropy}
\title{Compute the permutation entropy of a time series.}
\usage{
permutation_entropy(x)
}
\arguments{
\item{x}{A time series or vector.}
}
\value{
The permutation entropy of the time series.
}
\description{
Calculates the permutation entropy of a time 
 series using the pdc package. The miminimum
 entropy over the embedding dimensions 3:7 is
 returned.
}

