% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sample_entropy}
\alias{sample_entropy}
\title{Compute the sample entropy of the data}
\usage{
sample_entropy(x)
}
\arguments{
\item{x}{The data}
}
\value{
Sample entropy
}
\description{
Compute the sample entropy of the data
}
