% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{spectral_entropy}
\alias{spectral_entropy}
\title{Compute spectral entropy}
\usage{
spectral_entropy(x)
}
\arguments{
\item{x}{Time series}
}
\value{
return
}
\description{
Computes the entropy of the binned spectrogram.
 Wrapper of ForeCA package function.
}

